Crate junction_core

An XDS-compatible HTTP Client.

Structs

BackendLb

A Backend and the LoadBalancer it’s configured with.

Client

A service discovery client that looks up URL information based on URLs, headers, and methods.

Endpoint

An HTTP endpoint to make a request to.

HttpRequest

A view into an HTTP Request, before any rewrites or modifications have been made while sending or processing a response.

ResolvedRoute

The result of resolving a route.

ResourceVersion

An opaque string used to version an xDS resource.

Url

An [http::Uri] that has an http or https scheme and a non-empty authority.

XdsConfig

A single xDS configuration object, with additional metadata about when it was fetched and processed.

Enums

ConfigMode

How to resolve routes and endpoints.

EndpointAddress

The address of an endpoint.

Error

LoadBalancer

Functions

check_route

Check route resolution.

Type Aliases

Result

A Result alias where the Err case is junction_core::Error.