junction_core

Struct Client

pub struct Client { /* private fields */ }

A service discovery client that looks up URL information based on URLs, headers, and methods.

Clients use a shared in-memory cache to keep data warm so that a request never has to block on a remote service.

Clients are cheaply cloneable, and should be cloned to create multiple clients that share the same in-memory cache.

Implementations

impl Client

pub async fn build( address: String, node_id: String, cluster: String, ) -> Result<Self, Box<dyn Error>>

Build a new dynamic client, spawning a new ADS client in the background.

This method creates a new ADS client and ADS connection. Dynamic data will not be shared with existing clients. To create a client that shares data with existing clients, clone an existing client.

This function assumes that youโre currently running the context of a tokio runtime and spawns background work on a tokio executor.

pub fn with_static_config( self, routes: Vec<Route>, backends: Vec<Backend>, ) -> Client

Build a client with static configuration. This client will use the passed configuration to resolve routes and backends, but will still fetch endpoints dynamically.

This method will panic if the client being cloned is fully static. To convert a static client to a client that uses dynamic config, create a new client.

pub fn with_static_endpoints( self, routes: Vec<Route>, backends: Vec<Backend>, ) -> Client

Construct a client that uses fully static configuration and does not connect to a control plane at all.

This is intended to be used to test configuration in controlled settings or to use Junction an offline mode. Once a client has been converted to fully static, itโs not possible to convert it back to using dynamic discovery data.

pub async fn resolve_http( &self, method: &Method, url: &Url, headers: &HeaderMap, ) -> Result<Endpoint>

Resolve an HTTP method, URL, and headers into an Endpoint.

This is the main entry point into Junction. When building an integration, use this method to fetch an initial endpoint. After making an initial request, use report_status to report the status of the request and to retry on failure.

The endpoint returned from this method should be a complete description of how to make an HTTP request - it contains the IP address to use, the full URL and hostname, the complete set of headers, and retry and timeout policy the client should use to make a request.

pub async fn report_status( &self, endpoint: Endpoint, response: HttpResult, ) -> Result<Endpoint>

Report the status of an externally made HTTP request made against an Endpoint returned from resolve_http.

If retrying the response is appropriate, a new Endpoint will be returned with updated address and host info set - calling resolve_http to start a retry attempt will drop request history and may result in too many retries.

If a retry is not appropriate, the returned Endpoint will have updated history information, but request details will remain the same. Clients may use that value for status or error reporting.

pub async fn resolve_route( &self, request: HttpRequest<'_>, deadline: Option<Instant>, ) -> Result<ResolvedRoute>

Resolve an HTTP method, URL, and headers to a target backend, returning the Route that matched, the index of the rule that matched, and the backend that was chosen - to make backend choice determinstic with multiple backends, set the JUNCTION_SEED environment variable.

This is a lower-level method that only performs the Route matching part of resolution. Itโs intended for debugging or querying a client for specific information. For everyday use, prefer Client::resolve_http.

pub async fn select_endpoint( &self, backend: &BackendId, ctx: LbContext<'_>, deadline: Option<Instant>, ) -> Result<SelectedEndpoint>

Select an endpoint address for this backend from the set of currently available endpoints.

This is a lower level method that only performs part of route resolution, and is intended for debugging and testing. For everyday use, prefer Client::resolve_http.

pub async fn csds_server(self, port: u16) -> Result<(), Error>

Start a gRPC CSDS server on the given port. To run the server, you must await this future.

For static clients, this does nothing.

pub fn dump_xds(&self, not_found: bool) -> Vec<XdsConfig>

Dump the clientโs current cache of xDS resources, as fetched from the config server.

This is a programmatic view of the same data that you can fetch over gRPC by starting a Client::csds_server.

pub fn dump_xds_errors(&self) -> Vec<XdsConfig>

Dump xDS resources that failed to update. This is a view of the data returned by Client::dump_xds that only contains resources with errors.

pub fn dump_routes(&self) -> Vec<Arc<Route>>

Dump the Clientโs current table of [Route]s, merging together any default routes and remotely fetched routes the same way the client would when resolving endpoints.

pub fn dump_backends(&self) -> Vec<Arc<BackendLb>>

Dump the Clientโs current table of BackendLbs, merging together any default configuration and remotely fetched config the same way the client would when resolving endpoints.

pub fn dump_endpoints(&self, backend: &BackendId) -> Option<EndpointIter>

Return the endpoints currently in cache for this backend.

The returned endpoints are a snapshot of what is currently in cache and will not update as new discovery information is pushed.