Crate junction_core

An XDS-compatible HTTP Client.

Structs

BackendLb

A [Backend] and the LoadBalancer it’s configured with.

Client

A service discovery client that looks up URL information based on URLs, headers, and methods.

Endpoint

Error

An error when using the Junction client.

HttpRequest

An outgoing HTTP Request, before any rewrites or modifications have been made.

ResolvedRoute

The result of resolving a route (see Client::resolve_route).

ResourceVersion

An opaque string used to version an xDS resource.

Url

An [http::Uri] that has an http or https scheme and a non-empty authority.

XdsConfig

A single xDS configuration object, with additional metadata about when it was fetched and processed.

Enums

HttpResult

The result of making an HTTP request.

LoadBalancer

ResolveMode

The strategy for udpating Routes, Backends, and Endpoints while resolving.

Functions

check_route

Check route resolution.

Type Aliases

Result

A Result alias where the Err case is junction_core::Error.