Crate junction_core

The core implementation for Junction - an xDS dynamically-configurable API load-balancer library.

Getting Started

Structs

BackendLb

A [Backend][junction_api::backend::Backend] and the LoadBalancer it’s configured with.

Client

A service discovery client that looks up URL information based on URLs, headers, and methods.

Endpoint

Error

An error when using the Junction client.

HttpRequest

An outgoing HTTP Request, before any rewrites or modifications have been made.

LbContext

The context required to select an address from a backend. Includes the URL and headers from an outgoing request.

ResolvedRoute

The result of resolving a route (see Client::resolve_route).

ResourceVersion

An opaque string used to version an xDS resource.

SelectedEndpoint

The result of selecting an endpoint (see Client::select_endpoint).

Url

An Uri with an http or https scheme and a non-empty authority.

XdsConfig

A single xDS configuration object, with additional metadata about when it was fetched and processed.

Enums

HttpResult

The result of making an HTTP request.

LoadBalancer

Functions

check_route

Check route resolution.

Type Aliases

Result

A Result alias where the Err case is junction_core::Error.